backend/requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.7
dj-database-url==2.1.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-oauth-toolkit==1.7.1
django-ratelimit==4.1.0
django-axes==6.1.1
django-security==0.17.0
cryptography==41.0.7

# API Documentation
drf-spectacular==0.26.5

# Environment Variables
python-decouple==3.8

# Date/Time Handling
pytz==2023.3

# Image Processing (for QR codes)
Pillow==10.0.1
qrcode==7.4.2

# Notifications
twilio==8.9.1

# WebSocket Support
channels==4.0.0
channels-redis==4.1.0

# Cache & Redis
django-redis==5.4.0
redis==5.0.1

# Background Tasks
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.0

# Utilities
requests==2.31.0
python-dateutil==2.8.2

# Security & Monitoring
django-csp==3.7
sentry-sdk==1.38.0

# Development
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Testing
pytest==7.4.3
pytest-django==4.5.2
pytest-cov==4.1.0
factory-boy==3.3.0
faker==20.1.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0
uvicorn[standard]==0.24.0
